+++ js/util.js
@@ -1,0 +1,6 @@
+const getRandomInteger = (min, max) => Math.floor(Math.random() * (Math.floor(max) - Math.ceil(min) + 1) + Math.ceil(min));
+
+const getRandomArrayIndex = (elements) => getRandomInteger(elements.min, elements.max);
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger, getRandomArrayIndex, getRandomArrayElement};
